Add metadataBase to images, import Head component, update Job component, and refactor imports
