+++ app/(site)/about/page.tsx
@@ -2,7 +2,7 @@ import { getProfile, getJobs } from '@/sanity/sanity.query';
 import { PortableText } from "@portabletext/react";
 import Image from 'next/image';
 import { BiEnvelope, BiDownload, BiLinkExternal } from 'react-icons/bi';
-import Job from '../components/Job';
+import Job from '../../../components/Job';
 
 const About = async () => {
   const profile = await getProfile();
