+++ next.config.js
@@ -2,6 +2,7 @@
 const nextConfig = {
   images: {
     domains: ['cdn.sanity.io'],
+    metadataBase: 'https://christianonoh.vercel.app',
   },
 }
